GKI: msm: vidc: don't fail probe when sku_index != sku_version
The vidc sku version is determined based on an efuse. There are
2 versions: vidc0 and vidc1 in the DT. We need to make sure we don't
fail probing if the sku_index != sku_version so that the bus doesn't
remain on indefinitely. This change will detect the sku difference and
gracefully exit probing as if it was a success.

Bug: 158064806
Test: verify device enters suspend AOSD and CXSD
Change-Id: Ic6c03f72ef1589fc1e8b9cd1b7c4f9a802c5ad2f
Signed-off-by: Will McVicker <willmcvicker@google.com>
(cherry-picked from commit 86d99ba3e614753a2f4a8f335b045d8b661b5a3c)
Signed-off-by: Albert I <kras@raphielgang.org>
Signed-off-by: DarkAngelGR <asavvo01@gmail.com>